refactor(dspark): split DFlash base from DSpark subclass (vLLM-aligned)
Mirror upstream vLLM's qwen3_dflash/qwen3_dspark layering: Qwen3DFlashModel
is the block-diffusion backbone + feature-KV base (stages A/B/C), and
Qwen3DSparkModel extends it overriding only stage D with the low-rank Markov
transition-bias head.  Weight classes split the same way: Qwen3DFlashWeight
declares the shared backbone + fc + hidden_norm, Qwen3DSparkWeight adds the
markov_head.  DFlash vs DSpark is now class identity, not a runtime
has_markov_head ckpt probe.

Also clarify the aux_hidden_state_layer_ids -1 conversion in model_factory.py:
the speculators convention (embedding = implicit layer 0) lands on the same
physical layer as vLLM's dflash_config.target_layer_ids, with a note against
dropping the -1.
